+++ lib/criteria_selector.dart
@@ -13,7 +13,6 @@ export 'src/dropdown_selector_bar_theme.dart';
 export 'src/dropdown_selector_button.dart';
 export 'src/dropdown_selector_button_theme.dart';
 export 'src/dropdown_selector_controller.dart';
-export 'src/dropdown_selector_result.dart';
 export 'src/dropdown_tab_data.dart';
 export 'src/i18n/localizations_delegate.dart';
 export 'src/selector/constants.dart';